+++ src/components/scenes/LoginScene.tsx
@@ -186,7 +186,7 @@ export function LoginSceneComponent(props: Props) {
   // Wait for the sticky config to initialize before rendering anything
   useAsyncEffect(async () => {
     if (isMaestro()) return
-    setLegacyLanding((await getStickyConfigValue('legacyLanding')) === 'true')
+    setLegacyLanding((await getStickyConfigValue('legacyLanding')) === 'legacyLanding')
   }, [])
 
   return loggedIn ? (